+++ app/src/main/java/dev/leonlatsch/photok/ui/components/bindings/BindableBottomSheetDialogFragment.kt
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.ui.components
+package dev.leonlatsch.photok.ui.components.bindings
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -23,27 +23,23 @@ import android.view.ViewGroup
 import androidx.annotation.LayoutRes
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
-import androidx.fragment.app.Fragment
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 /**
- * Base for all fragments that use data binding.
+ * Base for all BottomSheetDialogs that use data binding.
  *
- * @param BindingType the binding type generated when adding <data> tag to a layout.
- * @param layout the layout id with the data binding.
+ * @param BindingType The type of the generated binding
+ * @param layout The layout resource id
  *
  * @since 1.0.0
  * @author Leon Latsch
  */
-abstract class BindableFragment<BindingType : ViewDataBinding>(
+abstract class BindableBottomSheetDialogFragment<BindingType : ViewDataBinding>(
     @LayoutRes private val layout: Int
-) : Fragment(), Bindable<BindingType> {
+) : BottomSheetDialogFragment(), Bindable<BindingType> {
 
     final override lateinit var binding: BindingType
 
-    /**
-     * Creates layout and binding.
-     * **ALWAYS** call super.onCreateView() when overwriting.
-     */
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
@@ -54,9 +50,6 @@ abstract class BindableFragment<BindingType : ViewDataBinding>(
         return binding.root
     }
 
-    /**
-     * Inserts the Bindings. Always call super.insertBindings() to set lifecycle owner.
-     */
     override fun bind(binding: BindingType) {
         binding.lifecycleOwner = this
     }
