+++ app/src/main/java/dev/leonlatsch/photok/ui/components/bindings/BindableFragment.kt
@@ -14,34 +14,49 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.ui.components
+package dev.leonlatsch.photok.ui.components.bindings
 
 import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import androidx.annotation.LayoutRes
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
-import dev.leonlatsch.photok.ui.components.base.BaseActivity
+import androidx.fragment.app.Fragment
 
 /**
- * Abstract class for Activities with binding context.
- * Inflates layout file and sets binding of type [BindingType].
+ * Base for all fragments that use data binding.
+ *
+ * @param BindingType the binding type generated when adding <data> tag to a layout.
+ * @param layout the layout id with the data binding.
  *
  * @since 1.0.0
  * @author Leon Latsch
  */
-abstract class BindableActivity<BindingType : ViewDataBinding>(
+abstract class BindableFragment<BindingType : ViewDataBinding>(
     @LayoutRes private val layout: Int
-) : BaseActivity(), Bindable<BindingType> {
+) : Fragment(), Bindable<BindingType> {
 
     final override lateinit var binding: BindingType
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = DataBindingUtil.setContentView(this, layout)
+    /**
+     * Creates layout and binding.
+     * **ALWAYS** call super.onCreateView() when overwriting.
+     */
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate(inflater, layout, container, false)
         bind(binding)
+        return binding.root
     }
 
+    /**
+     * Inserts the Bindings. Always call super.insertBindings() to set lifecycle owner.
+     */
     override fun bind(binding: BindingType) {
         binding.lifecycleOwner = this
     }
