+++ app/src/main/java/dev/leonlatsch/photok/ui/gallery/GalleryFragment.kt
@@ -39,8 +39,8 @@ import dev.leonlatsch.photok.databinding.FragmentGalleryBinding
 import dev.leonlatsch.photok.other.*
 import dev.leonlatsch.photok.ui.MainActivity
 import dev.leonlatsch.photok.ui.backup.ValidateBackupDialogFragment
-import dev.leonlatsch.photok.ui.components.BindableFragment
 import dev.leonlatsch.photok.ui.components.Dialogs
+import dev.leonlatsch.photok.ui.components.bindings.BindableFragment
 import dev.leonlatsch.photok.ui.process.DeleteBottomSheetDialogFragment
 import dev.leonlatsch.photok.ui.process.ExportBottomSheetDialogFragment
 import dev.leonlatsch.photok.ui.process.ImportBottomSheetDialogFragment
