+++ app/src/main/java/dev/leonlatsch/photok/ui/setup/SetupFragment.kt
@@ -30,9 +30,9 @@ import dev.leonlatsch.photok.other.empty
 import dev.leonlatsch.photok.other.getBaseApplication
 import dev.leonlatsch.photok.other.hide
 import dev.leonlatsch.photok.other.show
-import dev.leonlatsch.photok.ui.components.BindableFragment
 import dev.leonlatsch.photok.ui.components.Dialogs
 import dev.leonlatsch.photok.ui.components.base.BaseActivity
+import dev.leonlatsch.photok.ui.components.bindings.BindableFragment
 
 /**
  * Fragment for the setup.
