+++ app/src/main/java/dev/leonlatsch/photok/ui/unlock/UnlockFragment.kt
@@ -29,9 +29,9 @@ import dev.leonlatsch.photok.other.getBaseApplication
 import dev.leonlatsch.photok.other.hide
 import dev.leonlatsch.photok.other.show
 import dev.leonlatsch.photok.other.vanish
-import dev.leonlatsch.photok.ui.components.BindableFragment
 import dev.leonlatsch.photok.ui.components.Dialogs
 import dev.leonlatsch.photok.ui.components.base.BaseActivity
+import dev.leonlatsch.photok.ui.components.bindings.BindableFragment
 
 /**
  * Unlock fragment.
