+++ app/src/main/java/dev/leonlatsch/photok/ui/viewphoto/DetailsBottomSheetDialog.kt
@@ -21,7 +21,7 @@ import android.view.View
 import dev.leonlatsch.photok.R
 import dev.leonlatsch.photok.databinding.DialogBottomSheetDetailsBinding
 import dev.leonlatsch.photok.model.database.entity.Photo
-import dev.leonlatsch.photok.ui.components.BindableBottomSheetDialogFragment
+import dev.leonlatsch.photok.ui.components.bindings.BindableBottomSheetDialogFragment
 
 /**
  * Bottom Sheet Dialog for the photo details.
